made index tensor dtype long

+++ models/roomrecon_network.py
@@ -510,7 +510,7 @@ class RoomNet(nn.Module):
                 [label_target.shape[0], 4], device=label_target.device)
             for b in range(bs):
                 batch_ind = torch.nonzero(
-                    r_coords[:, -1] == b, as_tuple=False).squeeze(1)
+                    r_coords[:, -1] == b, as_tuple=False).squeeze(1).long()
                 anchors_target[batch_ind] = anchors_gt[b][label_target[batch_ind]]
                 residual_target[batch_ind] = residual_gt[b][label_target[batch_ind]]
                 planes_target[batch_ind] = planes_gt[b][label_target[batch_ind]]
